requirements-windows.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.34.0
openai==1.3.7
requests==2.31.0
python-dotenv==1.0.0
websockets==12.0
aiofiles==23.2.1
jinja2==3.1.2
httpx==0.25.2

# Optional packages for enhanced functionality
# Install these separately if needed:
# pip install numpy pandas --only-binary=all
# pip install --upgrade pip setuptools wheel